move functionality from BackgroundSyncService to BackgroundSyncReceiver
